fix(gestures): move page-jump bubble above tab footer for iOS touch fix
