+++ app/(tabs)/_layout.tsx
@@ -2,7 +2,9 @@ import { useCallback, useMemo } from "react";
 import { StyleSheet, View } from "react-native";
 import { Tabs, useRouter, useSegments } from "expo-router";
 
+import { PageJumpInput } from "../../src/components/page-jump-input";
 import { TabFooter, type TabId } from "../../src/components/tab-footer";
+import { useMushafStore } from "../../src/store/mushaf-store";
 import { useUiStore } from "../../src/store/ui-store";
 import { colors } from "../../src/theme";
 
@@ -16,6 +18,9 @@ export default function TabsLayout() {
   const segments = useSegments();
   const router = useRouter();
   const footerVisible = useUiStore((s) => s.footerVisible);
+  const currentPage = useMushafStore((s) => s.currentPage);
+  const readCount = useMushafStore((s) => s.readCount);
+  const setJumpToPage = useMushafStore((s) => s.setJumpToPage);
 
   const activeTab = useMemo<TabId>(() => {
     const tabSegment = segments[segments.length - 1];
@@ -30,10 +35,14 @@ export default function TabsLayout() {
     [activeTab, router],
   );
 
+  const handleJumpToPage = useCallback(
+    (page: number) => setJumpToPage(page),
+    [setJumpToPage],
+  );
+
   return (
     <View style={styles.container}>
       <View style={styles.content}>
-        {/* Keep both tabs mounted so tab switch and Continue Reading are instant (no remount). */}
         <Tabs
           detachInactiveScreens={false}
           screenOptions={{
@@ -59,6 +68,15 @@ export default function TabsLayout() {
           visible={footerVisible}
         />
       </View>
+      {activeTab === "mushaf" && (
+        <View pointerEvents="box-none" style={styles.pageJumpOverlay}>
+          <PageJumpInput
+            currentPage={currentPage}
+            onJumpToPage={handleJumpToPage}
+            readCount={readCount}
+          />
+        </View>
+      )}
     </View>
   );
 }
@@ -77,5 +95,8 @@ const styles = StyleSheet.create({
     right: 0,
     bottom: 0,
     overflow: "hidden",
+  },
+  pageJumpOverlay: {
+    ...StyleSheet.absoluteFillObject,
   },
 });
